+++ include/apppopups.h
@@ -3,6 +3,7 @@
 
 #include "imgui.h"
 #include "textscroller.h"
+#include "nxmp-gfx.h"
 
 extern int newResW;
 extern int newResH;
@@ -25,6 +26,36 @@ namespace Popups {
 		ImGui::EndPopup();
         
     };
+	
+	
+	inline void SetupNativePopup(const char *id) {
+        
+        ImGui::PushStyleVar(ImGuiStyleVar_WindowPadding, ImVec2(0, 0));
+		ImGui::PushStyleVar(ImGuiStyleVar_FramePadding, ImVec2(0, 0));
+		ImGui::PushStyleVar(ImGuiStyleVar_ItemSpacing, ImVec2(0, 0));
+		//ImGui::PushStyleVar(ImGuiStyleVar_FrameRounding, 5.0f);
+		ImGui::PushStyleVar(ImGuiStyleVar_WindowRounding, 6.0f);
+		
+		
+		ImGui::PushStyleColor(ImGuiCol_PopupBg,nxmpgfx::Popup_Bg_color);
+		ImGui::PushStyleColor(ImGuiCol_Button,nxmpgfx::Button_color);
+		ImGui::PushStyleColor(ImGuiCol_ButtonActive,nxmpgfx::ButtonActive_color);
+		ImGui::PushStyleColor(ImGuiCol_ButtonHovered,nxmpgfx::NavHover_color);
+		ImGui::PushStyleVar(ImGuiStyleVar_ButtonTextAlign, ImVec2(0.5f, 0.5f));
+		ImGui::PushStyleColor(ImGuiCol_HeaderHovered, nxmpgfx::NavHover_color);
+		ImGui::PushStyleColor(ImGuiCol_NavHighlight, nxmpgfx::Active_color);
+        ImGui::SetNextWindowPos(ImVec2(255.0f*multiplyRes, 214.0f*multiplyRes), ImGuiCond_Once);
+		ImGui::SetNextWindowSize(ImVec2(771.0f*multiplyRes, 292.0f*multiplyRes), ImGuiCond_Once);
+		ImGui::OpenPopup(id);
+	};
+    
+    inline void ExitNativePopup(void) {
+        ImGui::PopStyleVar(5);
+		ImGui::PopStyleColor(6);
+		ImGui::EndPopup();
+        
+    };
+	
 
 	void SaveSettingsPopup(void);
 	void ResumePopup(void);