+++ include/gui.h
@@ -55,7 +55,7 @@
 
 #include "touchcontrols.h"
 #include "shaderMania.h"
-#include "themes.h"
+
 #include "logger.h"